Move MessageCryptoAnnotations to a different package
