+++ app/core/src/main/java/com/fsck/k9/ui/crypto/MessageCryptoCallback.java
@@ -4,6 +4,8 @@ package com.fsck.k9.ui.crypto;
 import android.content.Intent;
 import android.content.IntentSender;
 
+import com.fsck.k9.mailstore.MessageCryptoAnnotations;
+
 
 public interface MessageCryptoCallback {
     void onCryptoHelperProgress(int current, int max);
